Adds PATCH to reconfigure
